Fix:MK-187
This fixed the heart icons' location on the regular discounts cards that are below the high discounts cards. The hearts used to be on the left. they are now positioned on the right.

This also fixed a weird behavior with the navigation of the carousel for these cards, where the cards will move differently than the carrousel on top.